Added two distinct functions for obtaining new bounds based on old for arbitrarily transformed parameters.
The first function maps the old parameters to new parameters 1:1 if f(Theta[i]) only depends on f(theta[i]). If f(Theta[i]) depends on f(theta[i], theta[j], i != j) a greedy heuristic is used to determine the combination of lower and upper bounds of theta to use, that give the minimum lower bound and maximum upper bound for the new parameter Theta.

The second function uses a NLP to solve for the minimum and maximum upper bounds for each new parameter in Theta. This method should be more robust than the greedy method, which I'm unsure will work if there are multiple local minima, which may be possible under some transforms. However, the NLP method may return values that are less exact than the greedy heuristic

This ensures that the bounds in the parameter space, match the bounds in old parameter space.