Added discussion of when the methods used by the transform bounds functions are valid.
Care is definitely needed when the transformation isn't monotonic (increasing or decreasing). E.g. if the mapping is Theta[i] = theta[i]^2 and the bounds for theta[i] cross zero (e.g. [-1, 1.05]), the first mapping function will produce bounds for the new variable of [1, 1.1025).

In contrast the NLP function can produce the correct equivalent mapping of [0.0, 1.1025], because it can use a proportion of lb[i] and ub[i] at the same time to minimise and/or maximise the new variable bounds. However, the NLP can get the wrong value for the max in this case if it starts at a bad position and goes to the wrong local maxima (i.e. 1.0 rather than 1.1025). Setting the start position at 0.5 (halfway between fully lower bound or fully upper bound) fixes it for this case, but will need to be explored to make sure it works for other cases