Corrected issues with surrogate.
However, I don't think it's suitable for use in my reports discussion as the parameterisation used to form the surrogate impacts the covariance matrix and resultantly impacts the width of the parameter confidence intervals and reference tolerance intervals (i.e. wider parameter range = wider intervals). This does not appear to be working quite how I'd expect/well enough to add it as an example